requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
asyncpg
passlib[argon2]
python-multipart
alembic
pydantic-settings
pytest
flake8
argon2-cffi
celery
redis
qdrant-client
langchain
langchain-community
pypdf
sentence-transformers
